Move obsolete files to the Obsolete folder
