+++ src/Hangfire.Core/Obsolete/IAuthorizationFilter.cs
@@ -1,5 +1,5 @@
 ﻿// This file is part of Hangfire.
-// Copyright © 2016 Sergey Odinokov.
+// Copyright © 2013-2014 Sergey Odinokov.
 // 
 // Hangfire is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Lesser General Public License as 
@@ -15,25 +15,15 @@
 // License along with Hangfire. If not, see <http://www.gnu.org/licenses/>.
 
 using System;
-using System.Threading.Tasks;
+using System.Collections.Generic;
 using Hangfire.Annotations;
 
+// ReSharper disable once CheckNamespace
 namespace Hangfire.Dashboard
 {
-    [Obsolete("Use IDashboardDispatcher-based dispatchers instead. Will be removed in 2.0.0.")]
-    public class RequestDispatcherWrapper : IDashboardDispatcher
+    [Obsolete("Please use `IDashboardAuthorizationFilter` instead. Will be removed in 2.0.0.")]
+    public interface IAuthorizationFilter
     {
-        private readonly IRequestDispatcher _dispatcher;
-        
-        public RequestDispatcherWrapper([NotNull] IRequestDispatcher dispatcher)
-        {
-            if (dispatcher == null) throw new ArgumentNullException(nameof(dispatcher));
-            _dispatcher = dispatcher;
-        }
-
-        public Task Dispatch(DashboardContext context)
-        {
-            return _dispatcher.Dispatch(RequestDispatcherContext.FromDashboardContext(context));
-        }
+        bool Authorize([NotNull] IDictionary<string, object> owinEnvironment);
     }
 }
