+++ src/Hangfire.Core/Obsolete/IRequestDispatcher.cs
@@ -14,17 +14,15 @@
 // You should have received a copy of the GNU Lesser General Public 
 // License along with Hangfire. If not, see <http://www.gnu.org/licenses/>.
 
-#if NETFULL
 using System;
-using System.Collections.Generic;
-using Hangfire.Annotations;
+using System.Threading.Tasks;
 
+// ReSharper disable once CheckNamespace
 namespace Hangfire.Dashboard
 {
-    [Obsolete("Please use `IDashboardAuthorizationFilter` instead. Will be removed in 2.0.0.")]
-    public interface IAuthorizationFilter
+    [Obsolete("Use the `IDashboardDispatcher` interface instead. Will be removed in 2.0.0.")]
+    public interface IRequestDispatcher
     {
-        bool Authorize([NotNull] IDictionary<string, object> owinEnvironment);
+        Task Dispatch(RequestDispatcherContext context);
     }
 }
-#endif
